[DEV-3724] moving psc object traversal logic to helper & using it in IDV model
